+++ ClearBank.DeveloperTest/PaymentProcessors/BacsPaymentProcessor.cs
@@ -1,17 +1,17 @@
 ﻿using ClearBank.DeveloperTest.Types;
 
-namespace ClearBank.DeveloperTest.Services.PaymentProcessors
+namespace ClearBank.DeveloperTest.PaymentProcessors
 {
-    public class ChapsPaymentProcessor : IPaymentProcessor
+    public class BacsPaymentProcessor : IPaymentProcessor
     {
         public bool IsMatch(PaymentScheme paymentScheme)
         {
-            return paymentScheme == PaymentScheme.Chaps;
+            return paymentScheme == PaymentScheme.Bacs;
         }
 
         public MakePaymentResult Process(Account account, decimal amount)
         {
-            if (account == null || !account.AllowedPaymentSchemes.HasFlag(AllowedPaymentSchemes.Chaps) || account.Status != AccountStatus.Live)
+            if (account == null || !account.AllowedPaymentSchemes.HasFlag(AllowedPaymentSchemes.Bacs))
                 return new MakePaymentResult { Success = false };
 
             return new MakePaymentResult { Success = true };
