+++ ClearBank.DeveloperTest/PaymentProcessors/ChapsPaymentProcessor.cs
@@ -1,17 +1,17 @@
 ﻿using ClearBank.DeveloperTest.Types;
 
-namespace ClearBank.DeveloperTest.Services.PaymentProcessors
+namespace ClearBank.DeveloperTest.PaymentProcessors
 {
-    public class FasterPaymentProcessor : IPaymentProcessor
+    public class ChapsPaymentProcessor : IPaymentProcessor
     {
         public bool IsMatch(PaymentScheme paymentScheme)
         {
-            return paymentScheme == PaymentScheme.FasterPayments;
+            return paymentScheme == PaymentScheme.Chaps;
         }
 
         public MakePaymentResult Process(Account account, decimal amount)
         {
-            if (account == null || !account.AllowedPaymentSchemes.HasFlag(AllowedPaymentSchemes.FasterPayments) || account.Balance < amount)
+            if (account == null || !account.AllowedPaymentSchemes.HasFlag(AllowedPaymentSchemes.Chaps) || account.Status != AccountStatus.Live)
                 return new MakePaymentResult { Success = false };
 
             return new MakePaymentResult { Success = true };
